+++ app/Http/Controllers/HashtagController.php
@@ -23,6 +23,7 @@ class HashtagController extends Controller
 
         return view('hashtags.index', [
             'hashtags' => $hashtags,
+            'trendingDays' => max(1, (int) config('openbook.hashtags.trending_days', 7)),
         ]);
     }